Add clear function and allow sketch
In createDefaultGrid function, added an eventListener for "mouseover" to
each gridSquare variable. If user hovers mouse over one of the squares
on the grid, then it permanently becomes black by adding a 'black' class.

Added an eventListener to the "clear" button that redirects it to the clearGrid
function. If user "clicks" clear then all the squares on the grid are turned back
to white. This is possible by removing 'black' class from all the grid squares.