Add new features
Added a coloring mode, shows when user is coloring and not coloring. (.mode)

Added an input and resize button which allows user to resize the grid between 2 - 100.
(nested under .changeGridContainer)

Added grid information such as what the current grid size is and if user resize was
successful or not. (.currentGridSize and .gridMessage above .changeGridContainer)