Finish product
Grid Information
Set default grid size to 16x16.

Added .addEventListener('click') to changeGridBtn. When user inputs val
and clicks resize button, checkGridResize function will be called. Inside
the function, the value input is checked to make sure it is between 2 and
100. If not, gridMessage writes an error on webpage. If yes, setGrid function
is called. Inside setGrid function, all gridItems(divs) are removed from
grid container. Then, the value specified is squared and looped through.
Throughout each iteration, a div element is appended to gridContainer and
an event lister for 'mouseover' is added to each one. In addition, the style
"grid-template-columns:" on gridContainer is set to "repeat('specified value', 1fr)"

Draw on grid
Added let variable 'click' and set to false. Added event listener 'click' to gridContainer.
If user clicks on anything inside of gridContainer, click will be set to
the opposite of what it currently is. (if currently true, then set to false
and vice verse) Going back to setGrid function inside of the loop, we added
event listener 'mouseover' to each gridSquare. This means if mouse is hovering
over any of the grid squares then the colorSquare function will be called. Inside colorSquare function,
if 'click' variable is true, then the color (black or rainbow) is chosen and coloring mode will be shown as
"on" on the webpage/turned to green. If 'click' is false, then coloring mode will be shown as
"off" on the webpage/turned to red.

Change color
Added let variable 'black' and set to true. Also added rainbowChoices, which has an array of
roy g. biv strings. Added event listener 'click' to blackColor and rainbowColor. If rainbowColor
is triggered, black is changed to false. If blackColor is triggered, black is changed to true.

Resetting grid
Added event listener 'click' to clearGrid which is linked up to a button on the
webpage. If user clicks, then resetGrid function is called. Inside the function,
all grid squares inside of the grid container have their background color set
back to white.